Add Flake lock files
This morning I initialised a template but Nix couldn't generate the lock
file because I had apparenly hit a GitHub API rate limit. I shouldn't
have to deal with that sort of issue. By including the lock file in the
template, I can avoid the GitHub API or work around it entirely.